bunch of misc bug fixes.
Fixes pokemobs sometimes going to sleep due to time during combat.
Fixes confusion and flinch attacks not registering.
Fixes a bunch of "move missed" messages not working.
Fixes tame pokemobs being used for interaction stuff still generating
berries magically.
Exposes pokemob inventory via the IItemHandler capability.

+++ Pokecube Core/src/main/java/pokecube/core/database/moves/MoveEntry.java
@@ -68,7 +68,7 @@ public class MoveEntry implements IMoveConstants
     public byte                    statusChange;
     public float                   statusChance;
     public byte                    change                   = CHANGE_NONE;
-    public int                     chanceChance             = 0;
+    public float                   chanceChance             = 0;
     public int[]                   attackerStatModification = { 0, 0, 0, 0, 0, 0, 0, 0 };
     public int                     attackerStatModProb      = 100;
     public int[]                   attackedStatModification = { 0, 0, 0, 0, 0, 0, 0, 0 };
